requirements.txt
PyYAML
jsonschema
netaddr
rfc3986
